Adjust myprofile position
