+++ package.json
@@ -19,6 +19,7 @@
     "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.1",
     "redux-logger": "^3.0.6",
+    "v8-compile-cache": "^2.3.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
