+++ src/components/MyProfile.jsx
@@ -1,39 +1,16 @@
 import "bootstrap/dist/css/bootstrap.min.css";
 import MissionProfile from "./Mission/MissionProfile";
 import { ReservedDragons } from "./Dragon";
-import { Table, Row, Col } from "react-bootstrap";
-import { useSelector } from "react-redux";
-
+import { ReservedRocket } from "./Rocket";
+import { Row, Col } from "react-bootstrap";
 
 const MyProfile = () => {
-  const { rockets } = useSelector((state) => state.rockets);
-  const filteredRockets = rockets.filter((rocket) => rocket.reserved);
   return (
     <>
       <Row className="m-5">
         <Col sm={4}>
           <h2>My Rockets</h2>
-          <tbody className="m-3">
-            <div>{filteredRockets >= 0 && <p>No rockets reserved</p>}</div>
-            {filteredRockets.map((rockets) => (
-              <Table>
-                <tr key={rockets.id}>
-                  <td
-                    hover
-                    variant="light"
-                    className="p-4"
-                    style={{
-                      width: 370,
-                      borderBlockColor: "#d1cece",
-                      borderWidth: 2,
-                    }}
-                  >
-                    {rockets.name}
-                  </td>
-                </tr>
-              </Table>
-            ))}
-          </tbody>
+          <ReservedRocket />
         </Col>
         <Col sm={4}>
           <div className="mission-profile">
